add 'TuringExt'; up deps
