+++ Project.toml
@@ -12,6 +12,8 @@ LatinHypercubeSampling = "a5e1c1ea-c99a-51d3-a14d-a9a37257b02d"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+
+[weakdeps]
 Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 
 [compat]
@@ -25,3 +27,6 @@ Optimization = "3.27, 4"
 Random = "1.10"
 Turing = "0.33.1, 0.34, 0.35"
 julia = "1.10"
+
+[extensions]
+TuringExt = "Turing"
