+++ examples/Project.toml
@@ -2,3 +2,5 @@
 BOSS = "15052f3f-a82a-4ca3-bbc4-f478a3bd1ae2"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 OptimizationPRIMA = "72f8369c-a2ea-4298-9126-56167ce9cbc2"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
