+++ test/Project.toml
@@ -1,3 +1,4 @@
 [deps]
 OptimizationPRIMA = "72f8369c-a2ea-4298-9126-56167ce9cbc2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
